+++ sql/create_databases.sql
@@ -11,7 +11,7 @@ CREATE DATABASE transaction_db;
 CREATE TABLE IF NOT EXISTS account_tab (
     id SERIAL PRIMARY KEY,
     account_id INT UNIQUE NOT NULL,
-    balance DECIMAL(20, 8) NOT NULL DEFAULT 0,
+    balance BIGINT NOT NULL DEFAULT 0,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
@@ -23,7 +23,7 @@ CREATE TABLE IF NOT EXISTS fund_movement_tab (
     stage INT NOT NULL,
     source_account_id INT NOT NULL,
     destination_account_id INT NOT NULL,
-    amount DECIMAL(20, 8) NOT NULL,
+    amount BIGINT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS transaction_tab (
     transaction_id CHAR(36) UNIQUE NOT NULL,
     source_account_id INT NOT NULL,
     destination_account_id INT NOT NULL,
-    amount DECIMAL(20, 8) NOT NULL,
+    amount BIGINT NOT NULL,
     transaction_status INT NOT NULL,
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
